Cleanup Filebeat publishers
* Separate async and sync publisher into own files
* Add Publish interface method which can be used to manually trigger publish (see https://github.com/elastic/beats/pull/2456)
* Add getDataEvents function
* Rename publish to publisher package
